Style nav, page header and main
