+++ src/js/app/index.js
@@ -1,64 +1,21 @@
-import React, { PureComponent } from 'react';
+import React from 'react';
 import { HashRouter, Route } from 'react-router-dom';
 import Navigation from './navigation';
-import PageTitle from './pages/page-title';
+import PageTitle from './page-title';
+import Scrollable from './scrollable';
 import routes from '../routes';
 
-class StringClass extends PureComponent {
+export default class App extends React.PureComponent {
   render () {
-    return 'String class';
-  }
-}
-
-class ArrayClass extends PureComponent {
-  render () {
-    return [
-      [
-        'Array',
-        'class'
-      ]
-    ];
-  }
-}
-
-const StringSFC = () => (
-  'String SFC'
-);
-
-const ArraySFC = () => (
-  [
-    'Array',
-    'SFC'
-  ]
-);
-
-export default class App extends PureComponent {
-  constructor (props) {
-    super(props);
-
-    this.state = {
-      value: ''
-    };
-
-    this.onChange = this.onChange.bind(this);
-  }
-
-  onChange (event) {
-    this.setState({
-      value: event.target.value
-    });
-  }
-
-  render () {
-    const { value } = this.state;
-
     return (
       <HashRouter>
         <table className="app">
           <tbody>
             <tr>
               <td className="navigation">
-                <Navigation />
+                <Scrollable>
+                  <Navigation />
+                </Scrollable>
               </td>
               <td className="main">
                 {routes.map(({path, name, component: PageComponent}) => (
@@ -71,62 +28,13 @@ export default class App extends PureComponent {
                         <PageTitle>
                           {name}
                         </PageTitle>
-                        <PageComponent />
+                        <Scrollable>
+                          <PageComponent />
+                        </Scrollable>
                       </div>
                     )}
                   />
                 ))}
-                <div>
-                  <p>
-                    Hello, World!
-                  </p>
-                  <p>
-                    <StringClass />
-                  </p>
-                  <p>
-                    <ArrayClass />
-                  </p>
-                  <p>
-                    <StringSFC />
-                  </p>
-                  <p>
-                    <ArraySFC />
-                  </p>
-                  <p>
-                    <input type="submit" value="Input type = submit" />
-                  </p>
-                  <p>
-                    <input type="button" value="Input type = button" />
-                  </p>
-                  <p>
-                    <input
-                      type="text"
-                      placeholder="Input type = text placeholder"
-                      value={value}
-                      onChange={this.onChange}
-                    />
-                  </p>
-                  <p>
-                    <img alt="Alt text" />
-                  </p>
-                  <p>
-                    <a href="#" title="Link title">
-                      Link with title
-                    </a>
-                  </p>
-                  <p>
-                    Alt number
-                    <img alt={1} />
-                  </p>
-                  <p>
-                    <a href="#" title={1}>
-                      Link with title that's a number
-                    </a>
-                  </p>
-                  <p>
-                    Short word that doesn't need scrambling: "hi"
-                  </p>
-                </div>
               </td>
             </tr>
           </tbody>
