+++ src/js/app/scrollable/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export default class Scrollable extends React.PureComponent {
+  render () {
+    const { children } = this.props;
+
+    return (
+      <div className="scrollable">
+        {children}
+      </div>
+    );
+  }
+}
