+++ src/less/index.less
@@ -1,3 +1,11 @@
+@background-color: #FFF;
+@font-color: #333;
+@margin-base: 10px;
+@padding-base: @margin-base;
+@title-font-size: 32px;
+@title-height: @title-font-size + @padding-base * 2;
+@shadow-color: rgba(0, 0, 0, 0.5);
+
 * {
   box-sizing: border-box;
 }
@@ -8,13 +16,19 @@ body {
   padding: 0;
   font-family: arial, helvetica, sans-serif;
   font-size: 16px;
-  color: #333;
-  background-color: #FFF;
+  color: @font-color;
+  background-color: @background-color;
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
+  overflow: hidden;
+}
+
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
 }
 
 .app {
@@ -25,18 +39,50 @@ body {
   height: 100%;
   table-layout: fixed;
 
-  td {
+  & > tbody > tr > td {
+    position: relative;
     vertical-align: top;
 
     &.main {
       vertical-align: middle;
-      position: relative;
     }
   }
 }
 
+.main .scrollable {
+  padding-top: @title-height;
+}
+
+.page-title {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  text-align: center;
+  z-index: 1;
+  margin: 0;
+  font-size: @title-font-size;
+  line-height: 1;
+  padding: @padding-base;
+  background: linear-gradient(@background-color, transparent);
+}
+
+.scrollable {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  z-index: 0;
+  padding: @padding-base;
+}
+
 .navigation {
+  background-color: @background-color;
+  box-shadow: 0 0 5px 0 @shadow-color;
   width: 220px;
+  z-index: 2;
 
   ul {
     list-style: none;
@@ -44,8 +90,8 @@ body {
     padding: 0;
 
     li {
-      margin: 10px 0;
-      padding: 0 10px;
+      margin: @margin-base 0;
+      padding: 0;
     }
   }
 }
